Final-Dominique Alexander
This website is to showcase all of the cool and interesting things I
have learned so far at Per Scholas. Each of the fonts used are free
fonts that can be found on dafont.com a site that host free fonts. I
used percentages for the majority of width and heights to make the page
scale down with style and grace. There is also a mobile menu available
through the website. When you decrease the browser windows width it
scales down the entire page into a miniature replica of the original
size. Making the site look exactly the same, but smaller for a mobile
phone or tablet. On the gallery page there is container of content
which is scrollable on the x-axis. Each image in the container was made
clickable. On the landing page, if you click anywhere on the screen you
can toggle the visibility and the sound of the background video. Also
on the “merch” page, short for merchandise, you can click on each item
which gives you a small description of the item in a modal. It is a
container for text or images and simulates a pop up effect. My cart
page is functioning as a calculator for all of the items, on the
“merch” page, are pre-loaded with the values of the prices for each
individual item. Essentially this type of cart would only work for a
few items because this cart is proportioned like a calculator now with
10 items but 40 would be terrible design in my opinion. However if I
wrap all the images in a div and make “overflow-y:auto;” to that div’s
id or class you could in theory have an infinite amount of items
because of making all the text that overflows on the y axis, stays
hidden, and scrollable at the same time.